fix non-determinism, more declarative approach
