+++ charts/Controls.tsx
@@ -457,7 +457,8 @@ class TimelineControl extends React.Component<{ chart: ChartConfig }> {
     render() {
         const { chart } = this.props
         if (chart.props.tab === "table") {
-            const years = chart.dataTableTransform.timelineYears
+            const { dataTableTransform } = chart
+            const years = dataTableTransform.timelineYears
             if (years.length === 0) {
                 return null
             }
@@ -465,7 +466,10 @@ class TimelineControl extends React.Component<{ chart: ChartConfig }> {
                 <Timeline
                     years={years}
                     onTargetChange={this.onChartTargetChange}
-                    startYear={chart.timeDomain[0]}
+                    startYear={
+                        dataTableTransform.autoSelectedStartYear ??
+                        chart.timeDomain[0]
+                    }
                     endYear={chart.timeDomain[1]}
                     onStartDrag={this.onTimelineStart}
                     onStopDrag={this.onTimelineStop}
